refactor: use FileUploader in the index.vue
